A lot more work on definition viewer
Definition viewer is now set up so that it is accessible from the view
of a record if the record displays a record of type metadata. It can be
opened from the tools menu with the new option "Show definition viewer",
and it opens in a managed gui by itself.

The presentation is now done using html lists.